Update list_item_crime.xml
The checkbox is causing the problem. It's focusable so it's handling your click instead of ignoring it and letting it be handled by the ListView. Now, your app crashes when a list item is clicked (so you'll have to sort that out next) but the CrimeListFragment.onListItemClick method is being called.